make identical datatype
https://lgtm.com/projects/g/TheAlgorithms/C/snapshot/4d0dbc401d67171d555d494c49a08e72303b5ba7/files/project_euler/problem_22/sol1.c?sort=name&dir=ASC&mode=heatmap#xa0d183509f9bbb35:1

+++ project_euler/problem_22/sol1.c
@@ -128,7 +128,7 @@ int main(int argc, char **argv)
     for (i = 0; i < COUNT; i++)
 #endif
     {
-        unsigned int score = 0;
+        long score = 0;
         /* score the alphabets in i^th name */
         for (int j = 0; names[i][j] != '\0'; j++)
             score += names[i][j] - 'A' +
